+++ app/src/main/resources/db-migration/00048__schema.sql
@@ -1,0 +1,364 @@
+update materials
+set maxCapacity = 150
+where materialName = 'Antimony';
+update materials
+set maxCapacity = 250
+where materialName = 'Arsenic';
+update materials
+set maxCapacity = 300
+where materialName = 'Basic Conductors';
+update materials
+set maxCapacity = 100
+where materialName = 'Biotech Conductors';
+update materials
+set maxCapacity = 200
+where materialName = 'Cadmium';
+update materials
+set maxCapacity = 300
+where materialName = 'Carbon';
+update materials
+set maxCapacity = 200
+where materialName = 'Chemical Distillery';
+update materials
+set maxCapacity = 150
+where materialName = 'Chemical Manipulators';
+update materials
+set maxCapacity = 300
+where materialName = 'Chemical Storage Units';
+update materials
+set maxCapacity = 250
+where materialName = 'Chromium';
+update materials
+set maxCapacity = 300
+where materialName = 'Compact Composites';
+update materials
+set maxCapacity = 150
+where materialName = 'Compound Shielding';
+update materials
+set maxCapacity = 200
+where materialName = 'Conductive Ceramics';
+update materials
+set maxCapacity = 150
+where materialName = 'Conductive Polymers';
+update materials
+set maxCapacity = 150
+where materialName = 'Configurable Components';
+update materials
+set maxCapacity = 150
+where materialName = 'Core Dynamics Composites';
+update materials
+set maxCapacity = 300
+where materialName = 'Crystal Shards';
+update materials
+set maxCapacity = 200
+where materialName = 'Electrochemical Arrays';
+update materials
+set maxCapacity = 100
+where materialName = 'Exquisite Focus Crystals';
+update materials
+set maxCapacity = 250
+where materialName = 'Filament Composites';
+update materials
+set maxCapacity = 250
+where materialName = 'Flawed Focus Crystals';
+update materials
+set maxCapacity = 200
+where materialName = 'Focus Crystals';
+update materials
+set maxCapacity = 200
+where materialName = 'Galvanising Alloys';
+update materials
+set maxCapacity = 250
+where materialName = 'Germanium';
+update materials
+set maxCapacity = 300
+where materialName = 'Grid Resistors';
+update materials
+set maxCapacity = 300
+where materialName = 'Guardian Power Cell';
+update materials
+set maxCapacity = 250
+where materialName = 'Guardian Power Conduit';
+update materials
+set maxCapacity = 200
+where materialName = 'Guardian Sentinel Weapon Parts';
+update materials
+set maxCapacity = 300
+where materialName = 'Guardian Technology Component';
+update materials
+set maxCapacity = 300
+where materialName = 'Guardian Wreckage Component';
+update materials
+set maxCapacity = 300
+where materialName = 'Heat Conduction Wiring';
+update materials
+set maxCapacity = 250
+where materialName = 'Heat Dispersion Plate';
+update materials
+set maxCapacity = 200
+where materialName = 'Heat Exchangers';
+update materials
+set maxCapacity = 250
+where materialName = 'Heat Resistant Ceramics';
+update materials
+set maxCapacity = 150
+where materialName = 'Heat Vanes';
+update materials
+set maxCapacity = 200
+where materialName = 'High Density Composites';
+update materials
+set maxCapacity = 250
+where materialName = 'Hybrid Capacitors';
+update materials
+set maxCapacity = 100
+where materialName = 'Imperial Shielding';
+update materials
+set maxCapacity = 100
+where materialName = 'Improvised Components';
+update materials
+set maxCapacity = 300
+where materialName = 'Iron';
+update materials
+set maxCapacity = 300
+where materialName = 'Lead';
+update materials
+set maxCapacity = 250
+where materialName = 'Manganese';
+update materials
+set maxCapacity = 200
+where materialName = 'Mechanical Components';
+update materials
+set maxCapacity = 250
+where materialName = 'Mechanical Equipment';
+update materials
+set maxCapacity = 300
+where materialName = 'Mechanical Scrap';
+update materials
+set maxCapacity = 200
+where materialName = 'Mercury';
+update materials
+set maxCapacity = 100
+where materialName = 'Military Grade Alloys';
+update materials
+set maxCapacity = 100
+where materialName = 'Military Supercapacitors';
+update materials
+set maxCapacity = 200
+where materialName = 'Molybdenum';
+update materials
+set maxCapacity = 300
+where materialName = 'Nickel';
+update materials
+set maxCapacity = 200
+where materialName = 'Niobium';
+update materials
+set maxCapacity = 100
+where materialName = 'Pharmaceutical Isolators';
+update materials
+set maxCapacity = 200
+where materialName = 'Phase Alloys';
+update materials
+set maxCapacity = 300
+where materialName = 'Phosphorus';
+update materials
+set maxCapacity = 150
+where materialName = 'Polonium';
+update materials
+set maxCapacity = 150
+where materialName = 'Polymer Capacitors';
+update materials
+set maxCapacity = 200
+where materialName = 'Precipitated Alloys';
+update materials
+set maxCapacity = 150
+where materialName = 'Proprietary Composites';
+update materials
+set maxCapacity = 100
+where materialName = 'Photo Heat Radiators';
+update materials
+set maxCapacity = 150
+where materialName = 'Proto light Alloys';
+update materials
+set maxCapacity = 100
+where materialName = 'Proto Radiolic Alloys';
+update materials
+set maxCapacity = 150
+where materialName = 'Refined Focus Crystals';
+update materials
+set maxCapacity = 300
+where materialName = 'Rhenium';
+update materials
+set maxCapacity = 150
+where materialName = 'Ruthenium';
+update materials
+set maxCapacity = 300
+where materialName = 'Salvage Alloys';
+update materials
+set maxCapacity = 150
+where materialName = 'Selenium';
+update materials
+set maxCapacity = 150
+where materialName = 'Sensor Fragment';
+update materials
+set maxCapacity = 250
+where materialName = 'Shield Emitters';
+update materials
+set maxCapacity = 200
+where materialName = 'Shield Sensors';
+update materials
+set maxCapacity = 300
+where materialName = 'Sulphur';
+update materials
+set maxCapacity = 150
+where materialName = 'Technetium';
+update materials
+set maxCapacity = 150
+where materialName = 'Tellurium';
+update materials
+set maxCapacity = 300
+where materialName = 'Tempered Alloys';
+update materials
+set maxCapacity = 150
+where materialName = 'Thermic Alloys';
+update materials
+set maxCapacity = 200
+where materialName = 'Tin';
+update materials
+set maxCapacity = 200
+where materialName = 'Tungsten';
+update materials
+set maxCapacity = 250
+where materialName = 'Vanadium';
+update materials
+set maxCapacity = 300
+where materialName = 'Worn Shield Emitters';
+update materials
+set maxCapacity = 150
+where materialName = 'Yttrium';
+update materials
+set maxCapacity = 250
+where materialName = 'Zinc';
+update materials
+set maxCapacity = 250
+where materialName = 'Zirconium';
+
+update materials
+set maxCapacity = 150
+where materialName = 'Aberrant Shield Pattern Analysis';
+update materials
+set maxCapacity = 100
+where materialName = 'Abnormal Compact Emissions Data';
+update materials
+set maxCapacity = 100
+where materialName = 'Adaptive Encryptors Capture';
+update materials
+set maxCapacity = 300
+where materialName = 'Anomalous Bulk Scan Data';
+update materials
+set maxCapacity = 250
+where materialName = 'Anomalous FSD Telemetry';
+update materials
+set maxCapacity = 300
+where materialName = 'Atypical Disrupted Wake Echoes';
+update materials
+set maxCapacity = 150
+where materialName = 'Atypical Encryption Archives';
+update materials
+set maxCapacity = 200
+where materialName = 'Classified Scan Databanks';
+update materials
+set maxCapacity = 100
+where materialName = 'Classified Scan Fragment';
+update materials
+set maxCapacity = 200
+where materialName = 'Cracked Industrial Firmware';
+update materials
+set maxCapacity = 100
+where materialName = 'Datamined Wake Exceptions';
+update materials
+set maxCapacity = 150
+where materialName = 'Decoded Emission Data';
+update materials
+set maxCapacity = 300
+where materialName = 'Distorted Shield Cycle Recordings';
+update materials
+set maxCapacity = 150
+where materialName = 'Divergent Scan Data';
+update materials
+set maxCapacity = 150
+where materialName = 'Eccentric Hyperspace Trajectories';
+update materials
+set maxCapacity = 300
+where materialName = 'Exceptional Scrambled Emission Data';
+update materials
+set maxCapacity = 150
+where materialName = 'Guardian Weapon Blueprint Fragment';
+update materials
+set maxCapacity = 250
+where materialName = 'Inconsistent Shield Soak Analysis';
+update materials
+set maxCapacity = 250
+where materialName = 'Irregular Emission Data';
+update materials
+set maxCapacity = 250
+where materialName = 'Modified Consumer Firmware';
+update materials
+set maxCapacity = 100
+where materialName = 'Modified Embedded Firmware';
+update materials
+set maxCapacity = 200
+where materialName = 'Open Symmetric Keys';
+update materials
+set maxCapacity = 150
+where materialName = 'Pattern Alpha Obelisk Data';
+update materials
+set maxCapacity = 150
+where materialName = 'Pattern Beta Obelisk Data';
+update materials
+set maxCapacity = 150
+where materialName = 'Pattern Delta Obelisk Data';
+update materials
+set maxCapacity = 150
+where materialName = 'Pattern Epsilon Obelisk Data';
+update materials
+set maxCapacity = 150
+where materialName = 'Pattern Gamma Obelisk Data';
+update materials
+set maxCapacity = 100
+where materialName = 'Peculiar Shield Frequency Data';
+update materials
+set maxCapacity = 150
+where materialName = 'Security Firmware Patch';
+update materials
+set maxCapacity = 300
+where materialName = 'Specialized Legacy Firmware';
+update materials
+set maxCapacity = 250
+where materialName = 'Strange Wake Solutions';
+update materials
+set maxCapacity = 250
+where materialName = 'Tagged Encryption Codes';
+update materials
+set maxCapacity = 200
+where materialName = 'Unexpected Emission Data';
+update materials
+set maxCapacity = 200
+where materialName = 'Unexpected Scan Archives';
+update materials
+set maxCapacity = 200
+where materialName = 'Untypical Shield Scans';
+update materials
+set maxCapacity = 300
+where materialName = 'Unusual Encrypted Files';
+update materials
+set maxCapacity = 250
+where materialName = 'Chemical Processors';
+update materials
+set maxCapacity = 250
+where materialName = 'Conductive Components';
+update materials
+set maxCapacity = 100
+where materialName = 'Proto Heat Radiators';
+update materials
+set maxCapacity = 200
+where materialName = 'Boron';